test(ffi): move double-free test to external file (Codacy new-unsafe gate)
The inline #[cfg(test)] fixture needed a raw-handle deref (unsafe { *args }),
which Codacy flags as new unsafe usage in src/. Move the test to
tests/double_free.rs (external test files are exempt); the guard in
wrap_native_fn is unchanged.
